URLパラメータを加工して DecomojiCollection 型に整形する処理を Collection ストアの action で行うようにした
